Extract archive_utils from run_batch_standardization
First slice of the Wave C decomposition. The recursive zip extraction
helpers (_extract_zip_files_recursive, _extract_archives_in_tree,
_should_skip_archive_member, _is_within_directory, _resolve_archive_max_depth)
move into a new self-contained scripts/archive_utils module along with
the suffix sets they consult (DATA_FILE_SUFFIXES, ARCHIVE_SUFFIXES,
MAPPING_SUFFIXES, DEFAULT_ARCHIVE_MAX_DEPTH).

The new module imports only from stdlib + scripts.convert_dv (for the
format-registry-derived suffix list, with the same defensive fallback as
before), so it can be imported and unit-tested without paying the cost
of pandas / transformers initialisation that a full
run_batch_standardization import incurs.

run_batch_standardization re-exports every name it previously owned so
external callers and the Wave B test suite (which imports the helpers
directly from scripts.run_batch_standardization) continue to work
unchanged. TABULAR_SUFFIXES is preserved as an alias for DATA_FILE_SUFFIXES.

Net diff in the orchestrator: -122 / +18 lines. Test result: 281 passed,
1 skipped (unchanged).
